split `fpy2.passes` into `fpy2.analysis` and `fpy2.transform`
