+++ fpy2/README.md
@@ -1,0 +1,10 @@
+## FPy submodules
+
+- `backend` - Code generation from FPyIR
+- `frontend` - Parsing / Checking / IR-codegen
+- `ir` - Intermediate representation
+- `number` - Numbers library
+- `passes` - Compiler passes
+- `runtime` - Runtime support
+- `typing` - Typing hints
+- `utils` - Utility functions
