+++ fpy2/passes/__init__.py
@@ -1,9 +1,9 @@
 """Analysis or transformation passes on the FPy IR."""
 
-from .define_use import DefineUse
-from .for_bundling import ForBundling
-from .func_update import FuncUpdate
-from .ssa import SSA
-from .simplify_if import SimplifyIf
-from .verify import VerifyIR
-from .while_bundling import WhileBundling
+from ..analysis.define_use import DefineUse
+from ..transform.for_bundling import ForBundling
+from ..transform.func_update import FuncUpdate
+from ..transform.ssa import SSA
+from ..transform.simplify_if import SimplifyIf
+from ..analysis.verify import VerifyIR
+from ..transform.while_bundling import WhileBundling
